Added nullable list tests
